Merge 2.3.0 into sync_master_230

+++ samples/client/petstore/tizen/src/Category.h
@@ -1,23 +1,7 @@
 /*
- * Copyright (c) 2016 Samsung Electronics Co., Ltd All Rights Reserved
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- * http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-/*
  * Category.h
  *
- * 
+ * A category for a pet
  */
 
 #ifndef _Category_H_
@@ -31,7 +15,7 @@ namespace Tizen {
 namespace ArtikCloud {
 
 
-/*! \brief 
+/*! \brief A category for a pet
  *
  */
 
@@ -41,7 +25,7 @@ public:
 	 */
 	Category();
 	Category(char* str);
-	
+
 	/*! \brief Destructor.
 	 */
 	virtual ~Category();
@@ -56,11 +40,11 @@ public:
 
 	/*! \brief Get 
 	 */
-	long getId();
+	long long getId();
 
 	/*! \brief Set 
 	 */
-	void setId(long  id);
+	void setId(long long  id);
 	/*! \brief Get 
 	 */
 	std::string getName();
@@ -70,11 +54,11 @@ public:
 	void setName(std::string  name);
 
 private:
-	long id;
+	long long id;
 	std::string name;
 	void __init();
 	void __cleanup();
-	
+
 };
 }
 }
