+++ samples/client/petstore/tizen/src/StoreManager.h
@@ -1,19 +1,3 @@
-/*
- * Copyright (c) 2016 Samsung Electronics Co., Ltd All Rights Reserved
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- * http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
 #ifndef _StoreManager_H_
 #define _StoreManager_H_
 
@@ -21,8 +5,8 @@
 #include <cstring>
 #include <list>
 #include <glib.h>
-#include <map>
 #include "Order.h"
+#include <map>
 #include "Error.h"
 
 namespace Tizen{
@@ -89,7 +73,7 @@ bool getInventoryAsync(char * accessToken,
  * \param userData The user data to be passed to the callback function.
  */
 bool getOrderByIdSync(char * accessToken,
-	long orderId, 
+	long long orderId, 
 	void(* handler)(Order, Error, void* )
 	, void* userData);
 
@@ -102,7 +86,7 @@ bool getOrderByIdSync(char * accessToken,
  * \param userData The user data to be passed to the callback function.
  */
 bool getOrderByIdAsync(char * accessToken,
-	long orderId, 
+	long long orderId, 
 	void(* handler)(Order, Error, void* )
 	, void* userData);
 /*! \brief Place an order for a pet. *Synchronous*
